+++ JuiceMaker+Combine/JuiceMaker+Combine/OrderViewController.swift
@@ -112,25 +112,3 @@ final class OrderViewController: UIViewController {
     }
     
 }
-
-extension UILabel {
-    convenience init(alignment: NSTextAlignment = .left, text: String? = nil) {
-        self.init()
-        self.textAlignment = alignment
-        self.text = text
-    }
-}
-
-extension UIButton {
-    convenience init(title: String? = nil) {
-        self.init()
-        self.setTitle(title, for: .normal)
-    }
-}
-
-extension UIStackView {
-    convenience init(arrangedSubviews : [UIView], axis: NSLayoutConstraint.Axis) {
-        self.init(arrangedSubviews: arrangedSubviews)
-        self.axis = axis
-    }
-}
